+++ Plutarch/Bool.hs
@@ -29,7 +29,7 @@ import Generics.SOP (
   ccompare_NS,
   hcliftA2,
  )
-import Plutarch.Internal (punsafeAsClosedTerm)
+import Plutarch.Internal (plet, punsafeAsClosedTerm)
 import Plutarch.Internal.Generic (PCode, PGeneric, gpfrom)
 import Plutarch.Internal.Other (
   DerivePNewtype,
@@ -45,6 +45,7 @@ import Plutarch.Internal.Other (
   pmatch,
   pto,
   (#),
+  (#$),
   type (:-->),
  )
 import Plutarch.Lift (
@@ -87,6 +88,13 @@ class PEq t => POrd t where
 
 infix 4 #<=
 infix 4 #<
+
+instance PEq PBool where
+  x #== y' = plet y' $ \y -> pif' # x # y #$ pnot # y
+
+instance POrd PBool where
+  x #< y = pif' # x # pconstant False # y
+  x #<= y = pif' # x # y # pconstant True
 
 instance PEq b => PEq (DerivePNewtype a b) where
   x #== y = pto x #== pto y