+++ CHANGELOG.md
@@ -101,6 +101,7 @@
   Module: `Plutarch.TryFrom`
 
   Added by: [#326](https://github.com/Plutonomicon/plutarch/pull/326)
+
 - `plutarch-extra`: Add a new directory scaffold "`plutarch-extra`" which will be home to everything too specific to not be in the
   main Plutarch repo. Also refactored the test library.
 
@@ -113,6 +114,14 @@
   Module: `Plutarch.PPrelude`
 
   Added by: [#356](https://github.com/Plutonomicon/plutarch/pull/356)
+
+- Add `PConstant` instance for `Maybe`, with corresponding `PLift` instance for `PMaybeData`.
+
+  Added by: [#371](https://github.com/Plutonomicon/plutarch/pull/371)
+
+- Add `POrd` and `PEq` derivation for data encoded types via `PIsDataReprInstances`.
+
+  Added by: [#371](https://github.com/Plutonomicon/plutarch/pull/371)
 
 # 1.1.0